New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: No              Module(s): pipeline            Description: Describe your changes here...
Data type for cluster_flags is changed from int to numpy.uint16 for 
saving memory usage.

	



git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project@25027 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ pipeline/pipeline/hsd/tasks/baseline/validation.py
@@ -533,7 +533,7 @@ class ValidateLineRaster(object):
         #     exceeded two (out of three) thresholds in smoothing, and
         #     exceeded three (out of four) thresholds in final.
         #
-        self.cluster_info['cluster_flag'] = numpy.zeros(GridCluster.shape, dtype=int)
+        self.cluster_info['cluster_flag'] = numpy.zeros(GridCluster.shape, dtype=numpy.uint16)
         threshold = [1.5, 0.5]
         self.__update_cluster_flag('detection', GridCluster, threshold, 1)
